workflow_test/pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["yenslife <77geo5rge6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.160.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
fastapi = "^0.115.8"
pydantic = "^2.10.6"
uvicorn = "^0.34.0"
rich = "^13.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
